requirements.txt
flask
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyotp
qrcode[pil]
